backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "markdown>=3.10",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "tiktoken>=0.12.0",
    "google-genai>=0.3.0",
    "pypdf>=5.1.0",
    "faiss-cpu>=1.9.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "numpy>=2.2.3",
]
